@@ -3,12 +3,12 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>com.xuxueli</groupId>
 	<artifactId>xxl-job</artifactId>
-	<version>2.1.0</version>
+	<version>2.3.1-SNAPSHOT</version>
 	<packaging>pom</packaging>
 
 	<name>${project.artifactId}</name>
-	<description>A lightweight distributed task scheduling framework.</description>
-	<url>http://www.xuxueli.com/</url>
+	<description>A distributed task scheduling framework.</description>
+	<url>https://www.xuxueli.com/</url>
 
 	<modules>
 		<module>xxl-job-core</module>
@@ -20,28 +20,29 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 		<maven.compiler.encoding>UTF-8</maven.compiler.encoding>
-		<maven.compiler.source>1.7</maven.compiler.source>
-		<maven.compiler.target>1.7</maven.compiler.target>
+		<maven.compiler.source>1.8</maven.compiler.source>
+		<maven.compiler.target>1.8</maven.compiler.target>
 		<maven.test.skip>true</maven.test.skip>
 
-		<xxl-rpc.version>1.4.1</xxl-rpc.version>
+		<netty-all.version>4.1.63.Final</netty-all.version>
+		<gson.version>2.8.6</gson.version>
 
-		<spring.version>4.3.24.RELEASE</spring.version>
-		<spring-boot.version>1.5.21.RELEASE</spring-boot.version>
-		<mybatis-spring-boot-starter.version>1.3.4</mybatis-spring-boot-starter.version>
-		<mysql-connector-java.version>5.1.47</mysql-connector-java.version>
+		<spring.version>5.3.6</spring.version>
+		<spring-boot.version>2.4.5</spring-boot.version>
 
-		<slf4j-api.version>1.7.26</slf4j-api.version>
-		<junit.version>4.12</junit.version>
+		<mybatis-spring-boot-starter.version>2.1.4</mybatis-spring-boot-starter.version>
+		<mysql-connector-java.version>8.0.24</mysql-connector-java.version>
 
-		<commons-exec.version>1.3</commons-exec.version>
+		<slf4j-api.version>1.7.30</slf4j-api.version>
+		<junit-jupiter.version>5.7.1</junit-jupiter.version>
+		<javax.annotation-api.version>1.3.2</javax.annotation-api.version>
 
-		<groovy.version>2.5.7</groovy.version>
+		<groovy.version>3.0.8</groovy.version>
 
-		<maven-source-plugin.version>3.1.0</maven-source-plugin.version>
-		<maven-javadoc-plugin.version>3.1.0</maven-javadoc-plugin.version>
+		<maven-source-plugin.version>3.2.1</maven-source-plugin.version>
+		<maven-javadoc-plugin.version>3.2.0</maven-javadoc-plugin.version>
 		<maven-gpg-plugin.version>1.6</maven-gpg-plugin.version>
-		<maven-war-plugin.version>3.2.2</maven-war-plugin.version>
+		<maven-war-plugin.version>3.3.1</maven-war-plugin.version>
 	</properties>
 
 	<build>
