@@ -1,18 +1,14 @@
 package com.xxl.job.admin.dao;
 
 import com.xxl.job.admin.core.model.XxlJobLog;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
 
 import javax.annotation.Resource;
 import java.util.Date;
 import java.util.List;
-import java.util.Map;
 
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(locations = "classpath*:spring/applicationcontext-*.xml")
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class XxlJobLogDaoTest {
 
     @Resource
@@ -27,7 +23,7 @@ public class XxlJobLogDaoTest {
         log.setJobGroup(1);
         log.setJobId(1);
 
-        int ret1 = xxlJobLogDao.save(log);
+        long ret1 = xxlJobLogDao.save(log);
         XxlJobLog dto = xxlJobLogDao.load(log.getId());
 
         log.setTriggerTime(new Date());
@@ -47,13 +43,10 @@ public class XxlJobLogDaoTest {
         dto = xxlJobLogDao.load(log.getId());
 
 
-        List<Map<String, Object>> list2 = xxlJobLogDao.triggerCountByDay(new Date(new Date().getTime() + 30*24*60*60*1000), new Date());
-
-        int ret4 = xxlJobLogDao.clearLog(1, 1, new Date(), 100);
+        List<Long> ret4 = xxlJobLogDao.findClearLogIds(1, 1, new Date(), 100, 100);
 
         int ret2 = xxlJobLogDao.delete(log.getJobId());
 
-        int ret3 = xxlJobLogDao.triggerCountByHandleCode(-1);
     }
 
 }
