@@ -1,17 +1,14 @@
 package com.xxl.job.executor.test;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest
 public class XxlJobExecutorExampleBootApplicationTests {
 
 	@Test
 	public void test() {
-
+		System.out.println(11);
 	}
 
 }
