@@ -1,34 +1,52 @@
 package com.xxl.job.admin.core.conf;
 
-import com.xxl.job.admin.dao.XxlJobGroupDao;
-import com.xxl.job.admin.dao.XxlJobInfoDao;
-import com.xxl.job.admin.dao.XxlJobLogDao;
-import com.xxl.job.admin.dao.XxlJobRegistryDao;
-import com.xxl.job.core.biz.AdminBiz;
+import com.xxl.job.admin.core.alarm.JobAlarmer;
+import com.xxl.job.admin.core.scheduler.XxlJobScheduler;
+import com.xxl.job.admin.dao.*;
+import org.springframework.beans.factory.DisposableBean;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Component;
 
 import javax.annotation.Resource;
 import javax.sql.DataSource;
+import java.util.Arrays;
 
 /**
  * xxl-job config
  *
  * @author xuxueli 2017-04-28
  */
-@Configuration
-public class XxlJobAdminConfig implements InitializingBean{
+
+@Component
+public class XxlJobAdminConfig implements InitializingBean, DisposableBean {
+
     private static XxlJobAdminConfig adminConfig = null;
     public static XxlJobAdminConfig getAdminConfig() {
         return adminConfig;
     }
 
+
+    // ---------------------- XxlJobScheduler ----------------------
+
+    private XxlJobScheduler xxlJobScheduler;
+
     @Override
     public void afterPropertiesSet() throws Exception {
         adminConfig = this;
+
+        xxlJobScheduler = new XxlJobScheduler();
+        xxlJobScheduler.init();
     }
+
+    @Override
+    public void destroy() throws Exception {
+        xxlJobScheduler.destroy();
+    }
+
+
+    // ---------------------- XxlJobScheduler ----------------------
 
     // conf
     @Value("${xxl.job.i18n}")
@@ -37,8 +55,17 @@ public class XxlJobAdminConfig implements InitializingBean{
     @Value("${xxl.job.accessToken}")
     private String accessToken;
 
-    @Value("${spring.mail.username}")
-    private String emailUserName;
+    @Value("${spring.mail.from}")
+    private String emailFrom;
+
+    @Value("${xxl.job.triggerpool.fast.max}")
+    private int triggerPoolFastMax;
+
+    @Value("${xxl.job.triggerpool.slow.max}")
+    private int triggerPoolSlowMax;
+
+    @Value("${xxl.job.logretentiondays}")
+    private int logretentiondays;
 
     // dao, service
 
@@ -51,14 +78,19 @@ public class XxlJobAdminConfig implements InitializingBean{
     @Resource
     private XxlJobGroupDao xxlJobGroupDao;
     @Resource
-    private AdminBiz adminBiz;
+    private XxlJobLogReportDao xxlJobLogReportDao;
     @Resource
     private JavaMailSender mailSender;
     @Resource
     private DataSource dataSource;
+    @Resource
+    private JobAlarmer jobAlarmer;
 
 
     public String getI18n() {
+        if (!Arrays.asList("zh_CN", "zh_TC", "en").contains(i18n)) {
+            return "zh_CN";
+        }
         return i18n;
     }
 
@@ -66,8 +98,29 @@ public class XxlJobAdminConfig implements InitializingBean{
         return accessToken;
     }
 
-    public String getEmailUserName() {
-        return emailUserName;
+    public String getEmailFrom() {
+        return emailFrom;
+    }
+
+    public int getTriggerPoolFastMax() {
+        if (triggerPoolFastMax < 200) {
+            return 200;
+        }
+        return triggerPoolFastMax;
+    }
+
+    public int getTriggerPoolSlowMax() {
+        if (triggerPoolSlowMax < 100) {
+            return 100;
+        }
+        return triggerPoolSlowMax;
+    }
+
+    public int getLogretentiondays() {
+        if (logretentiondays < 7) {
+            return -1;  // Limit greater than or equal to 7, otherwise close
+        }
+        return logretentiondays;
     }
 
     public XxlJobLogDao getXxlJobLogDao() {
@@ -86,8 +139,8 @@ public class XxlJobAdminConfig implements InitializingBean{
         return xxlJobGroupDao;
     }
 
-    public AdminBiz getAdminBiz() {
-        return adminBiz;
+    public XxlJobLogReportDao getXxlJobLogReportDao() {
+        return xxlJobLogReportDao;
     }
 
     public JavaMailSender getMailSender() {
@@ -96,6 +149,10 @@ public class XxlJobAdminConfig implements InitializingBean{
 
     public DataSource getDataSource() {
         return dataSource;
+    }
+
+    public JobAlarmer getJobAlarmer() {
+        return jobAlarmer;
     }
 
 }
