@@ -30,8 +30,7 @@ public class I18nUtil {
         try {
             // build i18n prop
             String i18n = XxlJobAdminConfig.getAdminConfig().getI18n();
-            i18n = (i18n!=null && i18n.trim().length()>0)?("_"+i18n):i18n;
-            String i18nFile = MessageFormat.format("i18n/message{0}.properties", i18n);
+            String i18nFile = MessageFormat.format("i18n/message_{0}.properties", i18n);
 
             // load prop
             Resource resource = new ClassPathResource(i18nFile);
