@@ -4,6 +4,7 @@ import com.xxl.job.admin.core.model.XxlJobRegistry;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -12,19 +13,25 @@ import java.util.List;
 @Mapper
 public interface XxlJobRegistryDao {
 
-    public int removeDead(@Param("timeout") int timeout);
+    public List<Integer> findDead(@Param("timeout") int timeout,
+                                  @Param("nowTime") Date nowTime);
 
-    public List<XxlJobRegistry> findAll(@Param("timeout") int timeout);
+    public int removeDead(@Param("ids") List<Integer> ids);
+
+    public List<XxlJobRegistry> findAll(@Param("timeout") int timeout,
+                                        @Param("nowTime") Date nowTime);
 
     public int registryUpdate(@Param("registryGroup") String registryGroup,
                               @Param("registryKey") String registryKey,
-                              @Param("registryValue") String registryValue);
+                              @Param("registryValue") String registryValue,
+                              @Param("updateTime") Date updateTime);
 
     public int registrySave(@Param("registryGroup") String registryGroup,
                             @Param("registryKey") String registryKey,
-                            @Param("registryValue") String registryValue);
+                            @Param("registryValue") String registryValue,
+                            @Param("updateTime") Date updateTime);
 
-    public int registryDelete(@Param("registryGroup") String registGroup,
+    public int registryDelete(@Param("registryGroup") String registryGroup,
                           @Param("registryKey") String registryKey,
                           @Param("registryValue") String registryValue);
 
