@@ -114,14 +114,14 @@ $(function() {
 					{ "data": 'jobGroup', "visible" : false},
 					{
 						"data": 'triggerTime',
-                        "width":'16%',
+                        "width":'20%',
 						"render": function ( data, type, row ) {
-							return data?moment(new Date(data)).format("YYYY-MM-DD HH:mm:ss"):"";
+							return data?moment(data).format("YYYY-MM-DD HH:mm:ss"):"";
 						}
 					},
 					{
 						"data": 'triggerCode',
-                        "width":'12%',
+                        "width":'10%',
 						"render": function ( data, type, row ) {
 							var html = data;
 							if (data == 200) {
@@ -136,21 +136,21 @@ $(function() {
 					},
 					{
 						"data": 'triggerMsg',
-                        "width":'12%',
+                        "width":'10%',
 						"render": function ( data, type, row ) {
 							return data?'<a class="logTips" href="javascript:;" >'+ I18n.system_show +'<span style="display:none;">'+ data +'</span></a>':I18n.system_empty;
 						}
 					},
 	                { 
 	                	"data": 'handleTime',
-                        "width":'16%',
+                        "width":'20%',
 	                	"render": function ( data, type, row ) {
-	                		return data?moment(new Date(data)).format("YYYY-MM-DD HH:mm:ss"):"";
+	                		return data?moment(data).format("YYYY-MM-DD HH:mm:ss"):"";
 	                	}
 	                },
 	                {
 						"data": 'handleCode',
-                        "width":'12%',
+                        "width":'10%',
 						"render": function ( data, type, row ) {
                             var html = data;
                             if (data == 200) {
@@ -167,7 +167,7 @@ $(function() {
 	                },
 	                { 
 	                	"data": 'handleMsg',
-                        "width":'12%',
+                        "width":'10%',
 	                	"render": function ( data, type, row ) {
 	                		return data?'<a class="logTips" href="javascript:;" >'+ I18n.system_show +'<span style="display:none;">'+ data +'</span></a>':I18n.system_empty;
 	                	}
@@ -180,11 +180,32 @@ $(function() {
 	                		// better support expression or string, not function
 	                		return function () {
 		                		if (row.triggerCode == 200 || row.handleCode != 0){
-		                			var temp = '<a href="javascript:;" class="logDetail" _id="'+ row.id +'">'+ I18n.joblog_rolling_log +'</a>';
+
+		                			/*var temp = '<a href="javascript:;" class="logDetail" _id="'+ row.id +'">'+ I18n.joblog_rolling_log +'</a>';
 		                			if(row.handleCode == 0){
 		                				temp += '<br><a href="javascript:;" class="logKill" _id="'+ row.id +'" style="color: red;" >'+ I18n.joblog_kill_log +'</a>';
-		                			}
-		                			return temp;
+		                			}*/
+		                			//return temp;
+
+									var logKillDiv = '';
+									if(row.handleCode == 0){
+										logKillDiv = '       <li class="divider"></li>\n' +
+											'       <li><a href="javascript:void(0);" class="logKill" _id="'+ row.id +'" >'+ I18n.joblog_kill_log +'</a></li>\n';
+									}
+
+									var html = '<div class="btn-group">\n' +
+										'     <button type="button" class="btn btn-primary btn-sm">'+ I18n.system_opt +'</button>\n' +
+										'     <button type="button" class="btn btn-primary btn-sm dropdown-toggle" data-toggle="dropdown">\n' +
+										'       <span class="caret"></span>\n' +
+										'       <span class="sr-only">Toggle Dropdown</span>\n' +
+										'     </button>\n' +
+										'     <ul class="dropdown-menu" role="menu" _id="'+ row.id +'" >\n' +
+										'       <li><a href="javascript:void(0);" class="logDetail" _id="'+ row.id +'" >'+ I18n.joblog_rolling_log +'</a></li>\n' +
+										logKillDiv +
+										'     </ul>\n' +
+										'   </div>';
+
+		                			return html;
 		                		}
 		                		return null;	
 	                		}
