@@ -1,7 +1,5 @@
 package com.xxl.job.core.handler;
 
-import com.xxl.job.core.biz.model.ReturnT;
-
 /**
  * job handler
  *
@@ -10,28 +8,21 @@ import com.xxl.job.core.biz.model.ReturnT;
 public abstract class IJobHandler {
 
 
-	/** success */
-	public static final ReturnT<String> SUCCESS = new ReturnT<String>(200, null);
-	/** fail */
-	public static final ReturnT<String> FAIL = new ReturnT<String>(500, null);
-	/** fail timeout */
-	public static final ReturnT<String> FAIL_TIMEOUT = new ReturnT<String>(502, null);
-
-
 	/**
 	 * execute handler, invoked when executor receives a scheduling request
 	 *
-	 * @param param
-	 * @return
 	 * @throws Exception
 	 */
-	public abstract ReturnT<String> execute(String param) throws Exception;
+	public abstract void execute() throws Exception;
 
+
+	/*@Deprecated
+	public abstract ReturnT<String> execute(String param) throws Exception;*/
 
 	/**
 	 * init handler, invoked when JobThread init
 	 */
-	public void init() {
+	public void init() throws Exception {
 		// do something
 	}
 
@@ -39,7 +30,7 @@ public abstract class IJobHandler {
 	/**
 	 * destroy handler, invoked when JobThread destroy
 	 */
-	public void destroy() {
+	public void destroy() throws Exception {
 		// do something
 	}
 
