@@ -84,8 +84,11 @@ public class FileUtil {
             in = new FileInputStream(file);
             in.read(filecontent);
             in.close();
+
+            return filecontent;
         } catch (Exception e) {
             logger.error(e.getMessage(), e);
+            return null;
         } finally {
             if (in != null) {
                 try {
@@ -95,7 +98,6 @@ public class FileUtil {
                 }
             }
         }
-        return filecontent;
     }
 
 
