@@ -19,17 +19,20 @@ public class XxlJobConfig {
     @Value("${xxl.job.admin.addresses}")
     private String adminAddresses;
 
+    @Value("${xxl.job.accessToken}")
+    private String accessToken;
+
     @Value("${xxl.job.executor.appname}")
-    private String appName;
+    private String appname;
+
+    @Value("${xxl.job.executor.address}")
+    private String address;
 
     @Value("${xxl.job.executor.ip}")
     private String ip;
 
     @Value("${xxl.job.executor.port}")
     private int port;
-
-    @Value("${xxl.job.accessToken}")
-    private String accessToken;
 
     @Value("${xxl.job.executor.logpath}")
     private String logPath;
@@ -38,12 +41,13 @@ public class XxlJobConfig {
     private int logRetentionDays;
 
 
-    @Bean(initMethod = "start", destroyMethod = "destroy")
+    @Bean
     public XxlJobSpringExecutor xxlJobExecutor() {
         logger.info(">>>>>>>>>>> xxl-job config init.");
         XxlJobSpringExecutor xxlJobSpringExecutor = new XxlJobSpringExecutor();
         xxlJobSpringExecutor.setAdminAddresses(adminAddresses);
-        xxlJobSpringExecutor.setAppName(appName);
+        xxlJobSpringExecutor.setAppname(appname);
+        xxlJobSpringExecutor.setAddress(address);
         xxlJobSpringExecutor.setIp(ip);
         xxlJobSpringExecutor.setPort(port);
         xxlJobSpringExecutor.setAccessToken(accessToken);
